Added Api Call To Twitch for timestamps of death at the streamtime edited the look of the forms Added Option to show the world death count as the global death count

+++ src/Controller/Forms/DeathController.cs
@@ -25,7 +25,7 @@ namespace DeathCounterHotkey.Controller.Forms
         {
             DeathModel deathModel = new DeathModel();
             deathModel.TimeStamp = DateTime.Now;
-            deathModel.StreamTime = _streamTimeController.GetTime();
+            deathModel.StreamTime = (int)_streamTimeController.GetTime();
             deathModel.LocationId = locationId;
 
             _context.Deaths.Add(deathModel);
